fixed DeleteBook request  and propper error message

+++ bookstore/src/app/user/profile/profile.component.ts
@@ -20,14 +20,22 @@ export class ProfileComponent implements OnInit {
     this.apiService.getBooks('/books.json').subscribe(
       (booksData: any) => {
         if (typeof booksData === 'object') {
-          this.userBooks = Object.values(booksData);
-          console.log(this.userBooks);
+          // Convert the booksData object to an array of books with their keys as bookId
+          this.userBooks = Object.keys(booksData).map(key => ({ ...booksData[key], bookId: key }));
+  
+          // Filter out books that do not belong to the logged-in user
+          const loggedInUserId = this.getLoggedInUserId();
+          if (loggedInUserId) {
+            this.userBooks = this.userBooks.filter(book => book.owner === loggedInUserId);
+          }
         } else {
           this.userBooks = booksData;
         }
       },
       (error) => {
         console.error('Error fetching user books:', error);
+        // If there's an error fetching books, set userBooks to an empty array
+        this.userBooks = [];
       }
     );
   }
@@ -35,12 +43,20 @@ export class ProfileComponent implements OnInit {
   getLoggedInUserId(): string | null {
     return localStorage.getItem('userId');
   }
-  deleteBook(bookId: string): void {
+
+  deleteBook(bookId: string | undefined ): void {
+    if (!bookId) {
+      console.error('Book ID is undefined.');
+      return;
+    }
     if (confirm('Are you sure you want to delete this book?')) {
       this.apiService.delete(`books/${bookId}.json`).subscribe(
         () => {
+         
+         
+          
           // Remove the deleted book from the userBooks array
-          this.userBooks = this.userBooks.filter((book) => book.owner !== bookId);
+          this.userBooks = this.userBooks.filter((book) => book.bookId !== bookId);
         },
         (error) => {
           console.error('Error deleting book:', error);
